Updated code modularity
